@@ -203,7 +203,7 @@ final class Version20231220130825 extends AbstractMigration
      *
      * @return string
      */
-    private static function getToOneQuery(string $jsonKey, string $parentTable, string $childTable, string $parentTableId = null, string $childTableId = 'id', bool $childHasRelations = true, bool $withWhereClause = true): string
+    private static function getToOneQuery(string $jsonKey, string $parentTable, string $childTable, ?string $parentTableId = null, string $childTableId = 'id', bool $childHasRelations = true, bool $withWhereClause = true): string
     {
         // Set the column name to use for "ON" in the Join clause. By default, the child table name with "_id" appended.
         // E.g. "UPDATE feed p INNER JOIN feed_source c ON p.feed_source_id = c.id"
