@@ -6,12 +6,11 @@ namespace App\Controller;
 
 use App\Repository\PlaylistScreenRegionRepository;
 use App\Utils\ValidationUtils;
-use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpKernel\Attribute\AsController;
 
 #[AsController]
-class PlaylistScreenRegionDeleteController extends AbstractController
+class PlaylistScreenRegionDeleteController extends AbstractTenantAwareController
 {
     public function __construct(
         private readonly PlaylistScreenRegionRepository $playlistScreenRegionRepository,
@@ -24,7 +23,7 @@ class PlaylistScreenRegionDeleteController extends AbstractController
         $regionUlid = $this->validationUtils->validateUlid($regionId);
         $playlistUlid = $this->validationUtils->validateUlid($playlistId);
 
-        $this->playlistScreenRegionRepository->deleteRelations($screenUlid, $regionUlid, $playlistUlid);
+        $this->playlistScreenRegionRepository->deleteRelations($screenUlid, $regionUlid, $playlistUlid, $this->getActiveTenant());
 
         return new JsonResponse(null, \Symfony\Component\HttpFoundation\Response::HTTP_NO_CONTENT);
     }
