@@ -8,13 +8,12 @@ use ApiPlatform\Metadata\Exception\InvalidArgumentException;
 use App\Repository\PlaylistScreenRegionRepository;
 use App\Utils\ValidationUtils;
 use Doctrine\Common\Collections\ArrayCollection;
-use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpKernel\Attribute\AsController;
 
 #[AsController]
-class PlaylistScreenRegionPutController extends AbstractController
+class PlaylistScreenRegionPutController extends AbstractTenantAwareController
 {
     public function __construct(
         private readonly PlaylistScreenRegionRepository $playlistScreenRegionRepository,
@@ -37,7 +36,7 @@ class PlaylistScreenRegionPutController extends AbstractController
         $collection = new ArrayCollection($content);
         $this->validate($collection);
 
-        $this->playlistScreenRegionRepository->updateRelations($screenUlid, $regionUlid, $collection);
+        $this->playlistScreenRegionRepository->updateRelations($screenUlid, $regionUlid, $collection, $this->getActiveTenant());
 
         return new JsonResponse(null, \Symfony\Component\HttpFoundation\Response::HTTP_CREATED);
     }
