@@ -6,12 +6,11 @@ namespace App\Controller;
 
 use App\Repository\PlaylistSlideRepository;
 use App\Utils\ValidationUtils;
-use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpKernel\Attribute\AsController;
 
 #[AsController]
-class PlaylistSlideDeleteController extends AbstractController
+class PlaylistSlideDeleteController extends AbstractTenantAwareController
 {
     public function __construct(
         private readonly PlaylistSlideRepository $playlistSlideRepository,
@@ -23,7 +22,7 @@ class PlaylistSlideDeleteController extends AbstractController
         $ulid = $this->validationUtils->validateUlid($id);
         $slideUlid = $this->validationUtils->validateUlid($slideId);
 
-        $this->playlistSlideRepository->deleteRelations($ulid, $slideUlid);
+        $this->playlistSlideRepository->deleteRelations($ulid, $slideUlid, $this->getActiveTenant());
 
         return new JsonResponse(null, \Symfony\Component\HttpFoundation\Response::HTTP_NO_CONTENT);
     }
