@@ -6,12 +6,11 @@ namespace App\Controller;
 
 use App\Repository\ScreenGroupCampaignRepository;
 use App\Utils\ValidationUtils;
-use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpKernel\Attribute\AsController;
 
 #[AsController]
-class ScreenGroupCampaignDeleteController extends AbstractController
+class ScreenGroupCampaignDeleteController extends AbstractTenantAwareController
 {
     public function __construct(
         private readonly ScreenGroupCampaignRepository $screenGroupCampaignRepository,
@@ -23,7 +22,7 @@ class ScreenGroupCampaignDeleteController extends AbstractController
         $ulid = $this->validationUtils->validateUlid($id);
         $campaignUlid = $this->validationUtils->validateUlid($campaignId);
 
-        $this->screenGroupCampaignRepository->deleteRelations($ulid, $campaignUlid);
+        $this->screenGroupCampaignRepository->deleteRelations($ulid, $campaignUlid, $this->getActiveTenant());
 
         return new JsonResponse(null, \Symfony\Component\HttpFoundation\Response::HTTP_NO_CONTENT);
     }
