@@ -8,13 +8,12 @@ use ApiPlatform\Metadata\Exception\InvalidArgumentException;
 use App\Repository\ScreenGroupCampaignRepository;
 use App\Utils\ValidationUtils;
 use Doctrine\Common\Collections\ArrayCollection;
-use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpKernel\Attribute\AsController;
 
 #[AsController]
-class ScreenGroupCampaignPutController extends AbstractController
+class ScreenGroupCampaignPutController extends AbstractTenantAwareController
 {
     public function __construct(
         private readonly ScreenGroupCampaignRepository $screenGroupCampaignRepository,
@@ -36,7 +35,7 @@ class ScreenGroupCampaignPutController extends AbstractController
         $collection = new ArrayCollection($content);
         $this->validate($collection);
 
-        $this->screenGroupCampaignRepository->updateRelations($ulid, $collection);
+        $this->screenGroupCampaignRepository->updateRelations($ulid, $collection, $this->getActiveTenant());
 
         return new JsonResponse(null, \Symfony\Component\HttpFoundation\Response::HTTP_CREATED);
     }
