@@ -11,6 +11,7 @@ use App\Service\UserService;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Attribute\AsController;
 use Symfony\Component\HttpKernel\Exception\BadRequestHttpException;
 use Symfony\Component\HttpKernel\Exception\ConflictHttpException;
@@ -25,8 +26,12 @@ class UserActivateController extends AbstractController
 
     public function __invoke(Request $request): JsonResponse
     {
-        $body = json_decode($request->getContent());
-        $activationCode = $body->activationCode;
+        $body = $request->toArray();
+        $activationCode = $body['activationCode'] ?? null;
+
+        if (null === $activationCode) {
+            throw new BadRequestHttpException('Missing activation code');
+        }
 
         try {
             $this->externalUserService->activateExternalUser($activationCode);
@@ -38,6 +43,6 @@ class UserActivateController extends AbstractController
             throw new ConflictHttpException($e->getMessage());
         }
 
-        return new JsonResponse(null, 204);
+        return new JsonResponse(null, Response::HTTP_NO_CONTENT);
     }
 }
