@@ -40,7 +40,7 @@ class UlidProvider extends Base
         parent::__construct($generator);
     }
 
-    public static function ulid(\DateTimeInterface $dateTime = null): Ulid
+    public static function ulid(?\DateTimeInterface $dateTime = null): Ulid
     {
         $mtime = $dateTime ? $dateTime->getTimestamp().'000' : null;
 
@@ -54,7 +54,7 @@ class UlidProvider extends Base
         return \DateTime::createFromImmutable($ulid->getDateTime());
     }
 
-    private static function doGenerate(string $mtime = null): string
+    private static function doGenerate(?string $mtime = null): string
     {
         $faker = Factory::create();
 
@@ -132,9 +132,6 @@ class UlidProvider extends Base
      * Function from Symfony\Component\Uid\BinaryUtil.
      *
      * @see https://github.com/symfony/uid/blob/5.3/Ulid.php
-     *
-     * @param string $digits
-     * @param array $map
      *
      * @return string
      */
