@@ -26,14 +26,9 @@ class Media extends AbstractTenantScopedEntity implements RelationsChecksumInter
 
     /**
      * @Vich\UploadableField(mapping="media_object", fileNameProperty="filePath")
-     *
-     * @Assert\File(
-     *     maxSize = "200000k",
-     *     mimeTypes = {"image/jpeg", "image/png","image/svg+xml", "video/webm", "video/mp4", "image/gif"},
-     *     mimeTypesMessage = "Please upload a valid image format: jpeg, svg, gif or png, or video format: webm or mp4"
-     * )
      */
     #[Vich\UploadableField(mapping: 'media_object', fileNameProperty: 'filePath', size: 'size')]
+    #[Assert\File(maxSize: '200000k', mimeTypes: ['image/jpeg', 'image/png', 'image/svg+xml', 'video/webm', 'video/mp4', 'image/gif'], mimeTypesMessage: 'Please upload a valid image format: jpeg, svg, gif or png, or video format: webm or mp4')]
     public ?File $file = null;
 
     #[ORM\Column(nullable: true)]
@@ -57,6 +52,9 @@ class Media extends AbstractTenantScopedEntity implements RelationsChecksumInter
     #[ORM\Column(type: \Doctrine\DBAL\Types\Types::STRING, options: ['default' => ''])]
     private string $sha = '';
 
+    /**
+     * @var \Doctrine\Common\Collections\Collection<int, \App\Entity\Tenant\Slide>
+     */
     #[ORM\ManyToMany(targetEntity: Slide::class, mappedBy: 'media')]
     private Collection $slides;
 
@@ -125,7 +123,7 @@ class Media extends AbstractTenantScopedEntity implements RelationsChecksumInter
         return $this;
     }
 
-    public function setFile(File $file = null): self
+    public function setFile(?File $file = null): self
     {
         $this->file = $file;
 
