@@ -53,9 +53,12 @@ class Screen extends AbstractTenantScopedEntity implements RelationsChecksumInte
      * @var \Doctrine\Common\Collections\Collection<int, \App\Entity\Tenant\PlaylistScreenRegion>|\App\Entity\Tenant\PlaylistScreenRegion[]
      */
     #[ORM\OneToMany(mappedBy: 'screen', targetEntity: PlaylistScreenRegion::class, orphanRemoval: true)]
-    #[ORM\OrderBy(['weight' => \Doctrine\Common\Collections\Criteria::ASC])]
+    #[ORM\OrderBy(['weight' => \Doctrine\Common\Collections\Order::Ascending->value])]
     private Collection $playlistScreenRegions;
 
+    /**
+     * @var \Doctrine\Common\Collections\Collection<int, \App\Entity\Tenant\ScreenGroup>
+     */
     #[ORM\ManyToMany(targetEntity: ScreenGroup::class, mappedBy: 'screens')]
     private Collection $screenGroups;
 
