@@ -264,19 +264,19 @@ class SparkleIOFeedType implements FeedTypeInterface
         // Collects trailing tags one by one.
         $trailingTags = [];
         $pattern = "/\s*#(?<tag>[^\s#]+)\n?$/u";
-        while (preg_match($pattern, $text, $matches)) {
+        while (preg_match($pattern, (string) $text, $matches)) {
             // We're getting tags in reverse order.
             array_unshift($trailingTags, $matches['tag']);
-            $text = preg_replace($pattern, '', $text);
+            $text = preg_replace($pattern, '', (string) $text);
         }
 
         // Wrap sections in p tags.
-        $text = preg_replace("/(.+)\n?/u", '<p>\1</p>', $text);
+        $text = preg_replace("/(.+)\n?/u", '<p>\1</p>', (string) $text);
 
         // Wrap inline tags.
         $pattern = '/(#(?<tag>[^\s#]+))/';
         $text = '<div class="text">'.preg_replace($pattern,
-            '<span class="tag">\1</span>', $text).'</div>';
+            '<span class="tag">\1</span>', (string) $text).'</div>';
         // Append tags.
         $text .= PHP_EOL.'<div class="tags">'.implode(' ',
             array_map(fn ($tag) => '<span class="tag">#'.$tag.'</span>', $trailingTags)).'</div>';
