@@ -23,7 +23,7 @@ class FeedRepository extends ServiceEntityRepository
         parent::__construct($registry, Feed::class);
     }
 
-    public function getById(Ulid $feedId): Querybuilder
+    public function getById(Ulid $feedId): QueryBuilder
     {
         $queryBuilder = $this->_em->createQueryBuilder();
         $queryBuilder->select('s')
