@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace App\Repository;
 
 use ApiPlatform\Metadata\Exception\InvalidArgumentException;
+use App\Entity\Tenant;
 use App\Entity\Tenant\Screen;
 use App\Entity\Tenant\ScreenGroup;
 use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
@@ -13,7 +14,6 @@ use Doctrine\ORM\EntityManagerInterface;
 use Doctrine\ORM\Query\Expr\Join;
 use Doctrine\ORM\QueryBuilder;
 use Doctrine\Persistence\ManagerRegistry;
-use Symfony\Bundle\SecurityBundle\Security;
 use Symfony\Component\Uid\Ulid;
 
 /**
@@ -25,14 +25,12 @@ use Symfony\Component\Uid\Ulid;
 class ScreenGroupRepository extends ServiceEntityRepository
 {
     private readonly EntityManagerInterface $entityManager;
-    private readonly Security $security;
 
-    public function __construct(ManagerRegistry $registry, Security $security)
+    public function __construct(ManagerRegistry $registry)
     {
         parent::__construct($registry, ScreenGroup::class);
 
         $this->entityManager = $this->getEntityManager();
-        $this->security = $security;
     }
 
     public function getScreenGroupsByScreenId(Ulid $screenUlid): QueryBuilder
@@ -44,10 +42,8 @@ class ScreenGroupRepository extends ServiceEntityRepository
         return $queryBuilder;
     }
 
-    public function updateRelations(Ulid $screenUlid, ArrayCollection $collection)
+    public function updateRelations(Ulid $screenUlid, ArrayCollection $collection, Tenant $tenant)
     {
-        $user = $this->security->getUser();
-        $tenant = $user->getActiveTenant();
         $screenRepos = $this->entityManager->getRepository(Screen::class);
         $screen = $screenRepos->findOneBy(['id' => $screenUlid, 'tenant' => $tenant]);
 
@@ -82,10 +78,8 @@ class ScreenGroupRepository extends ServiceEntityRepository
         }
     }
 
-    public function deleteRelations(Ulid $screenUlid, Ulid $screenGroupUlid)
+    public function deleteRelations(Ulid $screenUlid, Ulid $screenGroupUlid, Tenant $tenant)
     {
-        $user = $this->security->getUser();
-        $tenant = $user->getActiveTenant();
         $screenRepos = $this->entityManager->getRepository(Screen::class);
         $screen = $screenRepos->findOneBy(['id' => $screenUlid, 'tenant' => $tenant]);
         if (is_null($screen)) {
