@@ -22,8 +22,8 @@ use Symfony\Component\Uid\Ulid;
 
 class UserService
 {
-    public const EXTERNAL_USER_DEFAULT_NAME = 'EXTERNAL_NOT_SET';
-    public const CODE_ALPHABET = '0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZ';
+    final public const EXTERNAL_USER_DEFAULT_NAME = 'EXTERNAL_NOT_SET';
+    final public const CODE_ALPHABET = '0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZ';
 
     public function __construct(
         private readonly UserActivationCodeRepository $activationCodeRepository,
@@ -158,7 +158,7 @@ class UserService
         $bindKey = '';
 
         for ($i = 0; $i < $length; ++$i) {
-            $bindKey .= self::CODE_ALPHABET[rand(0, $charsLength - 1)];
+            $bindKey .= self::CODE_ALPHABET[random_int(0, $charsLength - 1)];
         }
 
         return $bindKey;
