@@ -6,12 +6,12 @@ namespace App\Utils;
 
 class Roles
 {
-    public const ROLE_USER = 'ROLE_USER';
-    public const ROLE_SCREEN = 'ROLE_SCREEN';
-    public const ROLE_EDITOR = 'ROLE_EDITOR';
-    public const ROLE_EXTERNAL_USER = 'ROLE_EXTERNAL_USER';
-    public const ROLE_EXTERNAL_USER_ADMIN = 'ROLE_EXTERNAL_USER_ADMIN';
-    public const ROLE_USER_ADMIN = 'ROLE_USER_ADMIN';
-    public const ROLE_ADMIN = 'ROLE_ADMIN';
-    public const ROLE_SUPER_ADMIN = 'ROLE_SUPER_ADMIN';
+    final public const ROLE_USER = 'ROLE_USER';
+    final public const ROLE_SCREEN = 'ROLE_SCREEN';
+    final public const ROLE_EDITOR = 'ROLE_EDITOR';
+    final public const ROLE_EXTERNAL_USER = 'ROLE_EXTERNAL_USER';
+    final public const ROLE_EXTERNAL_USER_ADMIN = 'ROLE_EXTERNAL_USER_ADMIN';
+    final public const ROLE_USER_ADMIN = 'ROLE_USER_ADMIN';
+    final public const ROLE_ADMIN = 'ROLE_ADMIN';
+    final public const ROLE_SUPER_ADMIN = 'ROLE_SUPER_ADMIN';
 }
